+++ agricultural_marketing/agricultural_marketing/report/items_list_grouped_by_supplier/items_list_grouped_by_supplier.js
@@ -1,0 +1,32 @@
+// Copyright (c) 2024, Muhammad Salama and contributors
+// For license information, please see license.txt
+
+frappe.query_reports["Items list grouped by supplier"] = {
+	"filters": [
+	    {
+           "fieldname": "supplier",
+           "fieldtype": "Link",
+           "label": "Supplier",
+           "options": "Supplier",
+           "wildcard_filter": 0
+        },
+        {
+           "fieldname": "item_code",
+           "fieldtype": "Data",
+           "label": "Item",
+           "wildcard_filter": 0
+        },
+        {
+           "fieldname": "from_date",
+           "fieldtype": "Date",
+           "label": "From Date",
+           "wildcard_filter": 0
+        },
+        {
+           "fieldname": "to_date",
+           "fieldtype": "Date",
+           "label": "To Date",
+           "wildcard_filter": 0
+        },
+	]
+};